+++ sopa/__init__.py
@@ -9,8 +9,6 @@ log = logging.getLogger("sopa")
 configure_logger(log)
 
 if not any(f"--{option}" in sys.argv for option in ["version", "help"]):  # no import for cli helpers
-    from spatialdata import read_zarr  # will set `dataframe.query-planning` to False
-
     from ._settings import settings
     from . import utils
     from . import shapes
@@ -18,5 +16,5 @@ if not any(f"--{option}" in sys.argv for option in ["version", "help"]):  # no i
     from . import spatial
     from . import segmentation
     from .aggregation import aggregate, overlay_segmentation
-    from .patches import make_transcript_patches, make_image_patches
+    from .patches import make_transcript_patches, make_image_patches, compute_embeddings, cluster_embeddings
     from .utils import get_spatial_image, get_spatial_element, to_intrinsic, get_boundaries