+++ sopa/io/explorer/points.py
@@ -5,6 +5,7 @@ from pathlib import Path
 import dask.dataframe as dd
 import numpy as np
 import zarr
+from zarr.storage import ZipStore
 
 from ._constants import ExplorerConstants, FileNames
 from .utils import explorer_file_path
@@ -12,9 +13,8 @@ from .utils import explorer_file_path
 log = logging.getLogger(__name__)
 
 
-def subsample_indices(n_samples, factor: int = 4):
-    n_sub = n_samples // factor
-    return np.random.choice(n_samples, n_sub, replace=False)
+def subsample_indices(indices: np.ndarray, factor: int = 4) -> np.ndarray:
+    return np.random.choice(indices, len(indices) // factor, replace=False)
 
 
 def write_transcripts(
@@ -44,28 +44,27 @@ def write_transcripts(
     grid_size = ExplorerConstants.GRID_SIZE / ExplorerConstants.PIXELS_TO_MICRONS * pixel_size
     df[gene] = df[gene].astype("category")
 
-    location = df[["x", "y"]]
-    location *= pixel_size
-    location = np.concatenate([location, np.zeros((num_transcripts, 1))], axis=1)
+    xy: np.ndarray = (df[["x", "y"]] * pixel_size).values
 
-    if location.min() < 0:
+    if xy.min() < 0:
         log.warning("Some transcripts are located outside of the image (pixels < 0)")
     log.info(f"Writing {len(df)} transcripts")
 
-    xmax, ymax = location[:, :2].max(axis=0)
+    xmax, ymax = xy.max(axis=0)
 
     gene_names = list(df[gene].cat.categories)
     num_genes = len(gene_names)
 
     codeword_gene_mapping = list(range(num_genes))
 
-    valid = np.ones((num_transcripts, 1))
-    uuid = np.stack([np.arange(num_transcripts), np.full(num_transcripts, 65535)], axis=1)
-    transcript_id = np.stack([np.arange(num_transcripts), np.full(num_transcripts, 65535)], axis=1)
-    gene_identity = df[gene].cat.codes.values[:, None]
-    codeword_identity = np.stack([gene_identity[:, 0], np.full(num_transcripts, 65535)], axis=1)
-    status = np.zeros((num_transcripts, 1))
-    quality_score = np.full((num_transcripts, 1), ExplorerConstants.QUALITY_SCORE)
+    valid = np.ones((num_transcripts, 1), dtype=np.uint8)
+    arange = np.arange(num_transcripts, dtype=np.uint32)
+    uuid = np.stack([arange, np.full(num_transcripts, 65535, dtype=np.uint32)], axis=1)
+    transcript_id = np.stack([arange, np.full(num_transcripts, 65535, dtype=np.uint32)], axis=1)
+    gene_identity = df[gene].cat.codes.values[:, None].astype(np.uint16)
+    codeword_identity = np.stack([gene_identity[:, 0], np.full(num_transcripts, 65535, dtype=np.uint16)], axis=1)
+    status = np.ones((num_transcripts, 1), dtype=np.uint8)
+    quality_score = np.full((num_transcripts, 1), ExplorerConstants.QUALITY_SCORE, dtype=np.float32)
 
     ATTRS = {
         "codeword_count": num_genes,
@@ -93,24 +92,55 @@ def write_transcripts(
         "grid_keys": [],
     }
 
-    with zarr.ZipStore(path, mode="w") as store:
-        g = zarr.group(store=store)
-        g.attrs.put(ATTRS)
+    subsampling_locs = {0: np.arange(num_transcripts)}
 
-        grids = g.create_group("grids")
+    for level in range(max_levels):
+        tile_size = grid_size * 2**level
+        level_xy = xy[subsampling_locs[level]]
 
-        for level in range(max_levels):
-            log.info(f"   > Level {level}: {len(location)} transcripts")
-            level_group = grids.create_group(level)
+        indices = np.floor(level_xy / tile_size).clip(0).astype(int)
+        tiles_str_indices = np.array([f"{tx},{ty}" for (tx, ty) in indices])
+
+        GRIDS_ATTRS["grid_array_shapes"].append([])
+        GRIDS_ATTRS["grid_number_objects"].append([])
+        GRIDS_ATTRS["grid_keys"].append([])
+
+        n_tiles_x, n_tiles_y = ceil(xmax / tile_size), ceil(ymax / tile_size)
+
+        for tx in range(n_tiles_x):
+            for ty in range(n_tiles_y):
+                str_index = f"{tx},{ty}"
+                loc = np.where(tiles_str_indices == str_index)[0]
+
+                n_points_tile = len(loc)
+
+                if n_points_tile == 0:
+                    continue
+
+                GRIDS_ATTRS["grid_array_shapes"][-1].append({})
+                GRIDS_ATTRS["grid_keys"][-1].append(str_index)
+                GRIDS_ATTRS["grid_number_objects"][-1].append(n_points_tile)
+
+        if n_tiles_x * n_tiles_y == 1:
+            GRIDS_ATTRS["number_levels"] = level + 1
+            break
+
+        if level + 1 < max_levels:
+            subsampling_locs[level + 1] = subsample_indices(subsampling_locs[level])
+
+    with ZipStore(path, mode="w") as store:
+        g = zarr.group(store=store, zarr_format=2, attributes=ATTRS)
+
+        grids = g.create_group("grids", attributes=GRIDS_ATTRS)
+
+        for level, level_locs in subsampling_locs.items():
+            log.info(f"   > Level {level}: {len(level_locs)} transcripts")
+            level_group = grids.create_group(str(level))
 
             tile_size = grid_size * 2**level
 
-            indices = np.floor(location[:, :2] / tile_size).clip(0).astype(int)
+            indices = np.floor(xy[level_locs] / tile_size).clip(0).astype(int)
             tiles_str_indices = np.array([f"{tx},{ty}" for (tx, ty) in indices])
-
-            GRIDS_ATTRS["grid_array_shapes"].append([])
-            GRIDS_ATTRS["grid_number_objects"].append([])
-            GRIDS_ATTRS["grid_keys"].append([])
 
             n_tiles_x, n_tiles_y = ceil(xmax / tile_size), ceil(ymax / tile_size)
 
@@ -125,73 +155,21 @@ def write_transcripts(
                     if n_points_tile == 0:
                         continue
 
-                    GRIDS_ATTRS["grid_array_shapes"][-1].append({})
-                    GRIDS_ATTRS["grid_keys"][-1].append(str_index)
-                    GRIDS_ATTRS["grid_number_objects"][-1].append(n_points_tile)
+                    location = np.concatenate([xy[level_locs][loc], _z(len(loc))], axis=1).astype(np.float32)
 
                     tile_group = level_group.create_group(str_index)
-                    tile_group.array(
-                        "valid",
-                        valid[loc],
-                        dtype="uint8",
-                        chunks=chunks,
-                    )
-                    tile_group.array(
-                        "status",
-                        status[loc],
-                        dtype="uint8",
-                        chunks=chunks,
-                    )
-                    tile_group.array(
-                        "location",
-                        location[loc],
-                        dtype="float32",
-                        chunks=chunks,
-                    )
-                    tile_group.array(
-                        "gene_identity",
-                        gene_identity[loc],
-                        dtype="uint16",
-                        chunks=chunks,
-                    )
-                    tile_group.array(
-                        "quality_score",
-                        quality_score[loc],
-                        dtype="float32",
-                        chunks=chunks,
-                    )
-                    tile_group.array(
-                        "codeword_identity",
-                        codeword_identity[loc],
-                        dtype="uint16",
-                        chunks=chunks,
-                    )
-                    tile_group.array(
-                        "uuid",
-                        uuid[loc],
-                        dtype="uint32",
-                        chunks=chunks,
-                    )
-                    tile_group.array(
-                        "id",
-                        transcript_id[loc],
-                        dtype="uint32",
-                        chunks=chunks,
-                    )
+                    tile_group.create_array("valid", data=valid[level_locs][loc], chunks=chunks)
+                    tile_group.create_array("status", data=status[level_locs][loc], chunks=chunks)
+                    tile_group.create_array("location", data=location, chunks=chunks)
+                    tile_group.create_array("gene_identity", data=gene_identity[level_locs][loc], chunks=chunks)
+                    tile_group.create_array("quality_score", data=quality_score[level_locs][loc], chunks=chunks)
+                    tile_group.create_array("codeword_identity", data=codeword_identity[level_locs][loc], chunks=chunks)
+                    tile_group.create_array("uuid", data=uuid[level_locs][loc], chunks=chunks)
+                    tile_group.create_array("id", data=transcript_id[level_locs][loc], chunks=chunks)
 
-            if n_tiles_x * n_tiles_y == 1:
-                GRIDS_ATTRS["number_levels"] = level + 1
-                break
 
-            sub_indices = subsample_indices(len(location))
-
-            location = location[sub_indices]
-            valid = valid[sub_indices]
-            status = status[sub_indices]
-            gene_identity = gene_identity[sub_indices]
-            quality_score = quality_score[sub_indices]
-            codeword_identity = codeword_identity[sub_indices]
-            uuid = uuid[sub_indices]
-            transcript_id = transcript_id[sub_indices]
-
-        grids.attrs.put(GRIDS_ATTRS)
+def _z(n: int) -> np.ndarray:
+    """Create a zero array with a small non-zero value at the first element to avoid reading issues."""
+    z = np.zeros((n, 1))
+    z[0] = 1e-4
+    return z